+++ src/core/index.ts
@@ -1,2 +1,0 @@
-export * from './providers/ProviderFactory';
-export * from './Logger';