+++ src/core/providers/strategies/LedgerProviderStrategy/LedgerProviderStrategy.ts
@@ -75,7 +75,7 @@ export class LedgerProviderStrategy extends BaseProviderStrategy {
     return this.buildProvider();
   };
 
-  private buildProvider = async () => {
+  private readonly buildProvider = async () => {
     if (!this.provider) {
       throw new Error(ProviderErrorsEnum.notInitialized);
     }
@@ -90,8 +90,16 @@ export class LedgerProviderStrategy extends BaseProviderStrategy {
     return provider;
   };
 
-  private ledgerLogin = async (
+  override login = async (
     options?: LoginOptionsTypes
+  ): Promise<{ address: string; signature: string }> => {
+    // we are no longer cancelling the login here, because cancelLogin already destroys the provider
+    const loginOptions = { ...options, shouldSkipCancelLogin: true };
+    return super.login(loginOptions);
+  };
+
+  private readonly ledgerLogin = async (
+    options?: LoginOptionsTypes & { addressIndex?: number }
   ): Promise<{ address: string; signature: string }> => {
     if (!this.provider) {
       throw new Error(ProviderErrorsEnum.notInitialized);
@@ -104,7 +112,7 @@ export class LedgerProviderStrategy extends BaseProviderStrategy {
     });
     return {
       address,
-      signature: signature || ''
+      signature: signature ?? ''
     };
   };
 
@@ -119,13 +127,13 @@ export class LedgerProviderStrategy extends BaseProviderStrategy {
       options,
       config: this.config!,
       manager: ledgerConnectManager,
-      provider: this.provider!,
+      provider: this.provider,
       eventBus: this.eventBus,
       login: this.ledgerLogin.bind(this)
     });
   };
 
-  private createEventBus = async (anchor?: HTMLElement) => {
+  private readonly createEventBus = async (anchor?: HTMLElement) => {
     const shouldInitiateLogin = !getIsLoggedIn();
 
     if (!shouldInitiateLogin) {
@@ -145,19 +153,22 @@ export class LedgerProviderStrategy extends BaseProviderStrategy {
     return this.eventBus;
   };
 
-  private signTransactions = async (transactions: Transaction[]) => {
+  private readonly signTransactions = async (transactions: Transaction[]) => {
     if (!this._signTransactions) {
       throw new Error(ProviderErrorsEnum.signTransactionsNotInitialized);
     }
 
+    await this.rebuildProvider();
+
     const signedTransactions = await signTransactions({
       transactions,
-      handleSign: this._signTransactions
+      handleSign: this._signTransactions.bind(this.provider)
     });
+
     return signedTransactions;
   };
 
-  private signMessage = async (message: Message): Promise<Message> => {
+  private readonly signMessage = async (message: Message): Promise<Message> => {
     if (!this.provider || !this._signMessage) {
       throw new Error(ProviderErrorsEnum.notInitialized);
     }
@@ -175,7 +186,7 @@ export class LedgerProviderStrategy extends BaseProviderStrategy {
   /**
    * Makes sure the device is accessible and if not, tries to initialize a new provider
    */
-  private rebuildProvider = async () => {
+  private readonly rebuildProvider = async () => {
     try {
       await this.provider?.getAddress(); // can communicate with device
     } catch (_err) {